Add isOpen state and handleTriggerClick function.
The isOpen state stores an object with Boolean values that represent
whether a dropdown menu is open or not.

The handleTriggerClick function is passed to an onChange event listener
on each dropdown menu trigger and flips its corresponding Boolean value
in the isOpen object.

The isOpen value of a dropdown menu is used to toggle the visibility of
that dropdown menu, set the aria-expanded attribute value, and rotate
the arrow of the dropdown trigger