Close dropdownMenu on click outside.
In the callback function of a useState hook, a click event listener was
added to the entire document.

Upon a click event, all quiz info dropdown menus are retrieved based on
their respective wrappers and iterated over.

If a wrapper does not contain the targeted element and the dropdown menu
in that wrapper is not hidden, the isOpen Boolean of that dropdown menu
is flipped.